+++ thread/linuxthread/clone/cloneuts.c
@@ -10,6 +10,8 @@
 #include <sys/utsname.h>
 #include <sys/wait.h>
 
+#include <inttypes.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -67,7 +69,7 @@ int main(int argc, char *argv[]) {
   pid = clone(childFunc, stackTop, CLONE_NEWUTS | SIGCHLD, argv[1]);
   if (pid == -1)
     errExit("clone");
-  printf("clone() returned %ld\n", (long)pid);
+  printf("clone() returned %" PRId32 "\n", (int32_t)pid);
 
   /* Parent falls through to here */
