+++ thread/pthread/piestimator/usinggsl/worker.h
@@ -2,16 +2,17 @@
 #define __PI_WORKER__
 
 #include <gsl/gsl_rng.h>
+#include <stdint.h>
 
 struct thread_info {
   pthread_t thread_id; /* ID returned by pthread_create() */
   int thread_num;      /* Application-defined thread # */
-  long long maxiter;
+  int64_t maxiter;
   int seedx;
   int seedy;
   gsl_rng *rngx;
   gsl_rng *rngy;
-  long long accepted;
+  int64_t accepted;
 };
 
 void *piworker(void *thread_num);
